slight refactor of where components are
