+++ packages/ui/pod-pcd-ui/src/shared.tsx
@@ -1,0 +1,7 @@
+import { styled } from "@pcd/passport-ui";
+
+export const Container = styled.div`
+  padding: 16px;
+  overflow: hidden;
+  width: 100%;
+`;
